feat: refactor contestant validation and error handling with shared constants
