+++ client/lib/validation/contestant.ts
@@ -1,27 +1,15 @@
 import { z } from "zod";
 
-export const contestantGenderValues = ["male", "female"] as const;
-
-export const contestantAcademicYearValues = [
-  "1st Year - 1st Semester",
-  "1st Year - 2nd Semester",
-  "2nd Year - 1st Semester",
-  "2nd Year - 2nd Semester",
-  "3rd Year - 1st Semester",
-  "3rd Year - 2nd Semester",
-  "4th Year - 1st Semester",
-  "4th Year - 2nd Semester",
-] as const;
-
-const birthdaySchema = z
-  .string()
-  .regex(/^\d{4}-\d{2}-\d{2}$/, "Birthday must be in YYYY-MM-DD format")
-  .optional();
+import {
+  contestantAcademicYearValues,
+  contestantGenderValues,
+  dateYyyyMmDdSchema,
+} from "@/lib/validation/shared";
 
 export const contestantFormSchema = z.object({
   name: z.string().trim().min(2, "Name is required"),
   nicOrStudentId: z.string().trim().min(1, "NIC or Student ID is required"),
-  birthday: birthdaySchema,
+  birthday: dateYyyyMmDdSchema.optional(),
   gender: z.enum(contestantGenderValues).optional(),
   academicYear: z.enum(contestantAcademicYearValues).optional(),
 });
